Replace fetch with typed API client
