--- Frontend/src/contexts/DataContext.js
@@ -1,207 +1,0 @@
-// DataContext.js
-import React, {
-  createContext,
-  useState,
-  useEffect,
-  useContext,
-  useCallback,
-} from "react";
-
-const DataContext = createContext();
-
-export const useData = () => useContext(DataContext);
-
-export const DataProvider = ({ children }) => {
-  const [ingredients, setIngredients] = useState([]);
-  const [ingredientsNeedsRefetch, setIngredientsNeedsRefetch] = useState(false);
-  const [possibleIngredientTags, setPossibleIngredientTags] = useState([]);
-  const [meals, setMeals] = useState([]);
-  const [possibleMealTags, setPossibleMealTags] = useState([]);
-  const [mealsNeedsRefetch, setMealsNeedsRefetch] = useState(false);
-  const [activeRequests, setActiveRequests] = useState(0);
-  const fetching = activeRequests > 0;
-
-  const startRequest = useCallback(() => {
-    setActiveRequests((prev) => prev + 1);
-  }, []);
-
-  const endRequest = useCallback(() => {
-    setActiveRequests((prev) => Math.max(prev - 1, 0));
-  }, []);
-
-  const ingredientProcessingTagNames = [
-    "Whole Food",
-    "Lightly Processed",
-    "Highly Processed",
-  ];
-  const ingredientGroupTagNames = [
-    "Vegetable",
-    "Fruit",
-    "Meat",
-    "Dairy",
-    "Grain",
-  ];
-
-  const ingredientProcessingTags = possibleIngredientTags
-    ? possibleIngredientTags.filter(({ name }) =>
-        ingredientProcessingTagNames.includes(name),
-      )
-    : [];
-  const ingredientGroupTags = possibleIngredientTags
-    ? possibleIngredientTags.filter(({ name }) =>
-        ingredientGroupTagNames.includes(name),
-      )
-    : [];
-  const ingredientOtherTags = possibleIngredientTags
-    ? possibleIngredientTags.filter(
-        ({ name }) =>
-          !ingredientProcessingTagNames.includes(name) &&
-          !ingredientGroupTagNames.includes(name),
-      )
-    : [];
-
-  const mealDietTagNames = ["Vegetarian", "Vegan", "Carnivore"];
-  const mealTypeTagNames = ["Breakfast", "Lunch", "Dinner", "Snack"];
-
-  const mealDietTags = possibleMealTags
-    ? possibleMealTags.filter(({ name }) => mealDietTagNames.includes(name))
-    : [];
-  const mealTypeTags = possibleMealTags
-    ? possibleMealTags.filter(({ name }) => mealTypeTagNames.includes(name))
-    : [];
-  const mealOtherTags = possibleMealTags
-    ? possibleMealTags.filter(
-        ({ name }) =>
-          !mealDietTagNames.includes(name) && !mealTypeTagNames.includes(name),
-      )
-    : [];
-
-  const fetchData = useCallback(
-    async (url, setData, errorHandler, processData) => {
-      startRequest();
-      try {
-        const response = await fetch(url);
-        if (!response.ok) {
-          throw new Error(`HTTP error! Status: ${response.status}`);
-        }
-        const data = await response.json();
-        const processedData = processData ? processData(data) : data;
-        setData(processedData);
-      } catch (error) {
-        console.error("Error fetching data:", error);
-        errorHandler(error);
-      } finally {
-        endRequest();
-      }
-    },
-    [startRequest, endRequest],
-  );
-
-  const fetchIngredients = useCallback(() => {
-    const url = "/api/ingredients";
-    const processData = (data) =>
-      data.map((ingredient) => {
-        const unitsWithFloatGrams = ingredient.units.map((unit) => ({
-          ...unit,
-          grams: parseFloat(unit.grams),
-        }));
-        const defaultUnit =
-          unitsWithFloatGrams.find((unit) => unit.name === "1g") ||
-          unitsWithFloatGrams.find((unit) => unit.grams === 1) ||
-          unitsWithFloatGrams[0];
-        return {
-          ...ingredient,
-          units: unitsWithFloatGrams,
-          selectedUnitId: defaultUnit ? defaultUnit.id : null,
-        };
-      });
-
-    fetchData(
-      url,
-      setIngredients,
-      () => setIngredientsNeedsRefetch(true),
-      processData,
-    );
-  }, [fetchData]);
-
-  const fetchPossibleIngredientTags = useCallback(() => {
-    const url = "/api/ingredients/possible_tags";
-    fetchData(url, setPossibleIngredientTags, () =>
-      console.error("Error fetching tags"),
-    );
-  }, [fetchData]);
-
-  const fetchMeals = useCallback(() => {
-    const url = "/api/meals";
-    const processData = (data) =>
-      data.map((meal) => ({
-        ...meal,
-        ingredients: meal.ingredients
-          ? meal.ingredients.map((mi) => ({
-              ...mi,
-              unit_quantity: mi.unit_quantity
-                ? parseFloat(mi.unit_quantity)
-                : 0,
-            }))
-          : [],
-      }));
-
-    fetchData(url, setMeals, () => setMealsNeedsRefetch(true), processData);
-  }, [fetchData]);
-
-  const fetchPossibleMealTags = useCallback(() => {
-    const url = "/api/meals/possible_tags";
-    fetchData(url, setPossibleMealTags, () =>
-      console.error("Error fetching tags"),
-    );
-  }, [fetchData]);
-
-  //#region Effects
-  useEffect(() => {
-    fetchIngredients();
-    fetchPossibleIngredientTags();
-    fetchMeals();
-    fetchPossibleMealTags();
-  }, [
-    fetchIngredients,
-    fetchPossibleIngredientTags,
-    fetchMeals,
-    fetchPossibleMealTags,
-  ]); // Initial fetch
-
-  useEffect(() => {
-    if (ingredientsNeedsRefetch) {
-      fetchIngredients();
-      setIngredientsNeedsRefetch(false);
-    }
-    if (mealsNeedsRefetch) {
-      fetchMeals();
-      setMealsNeedsRefetch(false);
-    }
-  }, [
-    ingredientsNeedsRefetch,
-    mealsNeedsRefetch,
-    fetchIngredients,
-    fetchMeals,
-  ]); // Handle needsRefetch
-  //#endregion Effects
-
-  const value = {
-    ingredients,
-    setIngredients,
-    ingredientProcessingTags,
-    ingredientGroupTags,
-    ingredientOtherTags,
-    meals,
-    mealDietTags,
-    mealTypeTags,
-    mealOtherTags,
-    setIngredientsNeedsRefetch,
-    setMealsNeedsRefetch,
-    fetching,
-    startRequest,
-    endRequest,
-  };
-
-  return <DataContext.Provider value={value}>{children}</DataContext.Provider>;
-};
